New preprocessor

+++ compile.py
@@ -9,14 +9,14 @@
 #   
 #
 import time
+import traceback
 import os
 import argparse as arg
 from Classes.Error import *
 from globals import *
 import cProfile, pstats, io
 
-
-from PreParser import PreParser
+from Lexer import Lexer
 from PreParser import PreProcessor
 from Compiler import Compiler
 from Linker import *
@@ -29,25 +29,32 @@ def main():
         raw = inpf.read().decode()
 
     # preprocess
-    pre = PreParser(raw,config.__fileinput__)
-    pretokens = pre.getTokens() 
+    lex = Lexer(config.__fileinput__, raw)
+    firstTokens = lex.getTokens()
 
-    pp = PreProcessor(raw,pretokens,config.__fileinput__)
+
+    pp = PreProcessor(firstTokens)
     totals = pp.process()
-    #print("+-+-+ Compile +-+-+")
+    #print(totals)
+    #
+
     
     # global compilation
     c = Compiler()
     config.GlobalCompiler = c
     
-    
-    for file in totals:
-        
-        config.raw_filedata.append(file)
-        c.compile(file)
+
+    c.compile(totals)
 
     # function compilation
-    c.finalize()
+    
+    try:
+        c.finalize()
+    except AttributeError as e:
+        traceback.print_tb(e.__traceback__)
+        print(e)
+        print(c.currentfunction.current_token.start)
+        exit(1)
 
     if(c.panicmode):
         print("Could not finish compilation due to errors.")
@@ -76,8 +83,9 @@ def main():
         f.write(asm.encode())
 
     os.system(assemble(config.__fileoutput__))
-    os.system(link(config.__fileoutput__,config.__fileoutput__))
-    os.remove(config.__fileoutput__+".o")
+    if(config.__executable__):
+        os.system(link(config.__fileoutput__,config.__fileoutput__))
+        os.remove(config.__fileoutput__+".o")
 
     if(not config.__tonasm__):
         os.remove(config.__fileoutput__+".asm")
